changing scall to float

+++ wireframe/ft_scale_line.c
@@ -6,13 +6,13 @@
 /*   By: amersoul <amersoul@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/10/22 19:46:48 by amersoul          #+#    #+#             */
-/*   Updated: 2018/10/24 11:17:53 by amersoul         ###   ########.fr       */
+/*   Updated: 2018/10/24 15:34:40 by amersoul         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "wireframe.h"
 
-t_edge	ft_scale_edge(t_edge edge, const int scale)
+t_edge	ft_scale_edge(t_edge edge, const float scale)
 {
 	edge.vertex_1.x *= scale;
 	edge.vertex_1.y *= scale;
